requirements.txt
lxml==4.9.3
numpy==1.24.4
onnxruntime==1.16.3
paddle==1.0.2
Pillow==10.1.0
Pillow==10.1.0
Requests==2.31.0
selenium==4.16.0
torch==2.1.1
torchvision==0.16.1
tqdm==4.66.1
x2paddle==1.4.1
